Groups other navigation related abstractions to separate files and folders
