+++ workspaces/homepage/src/components/navigation/Navigation.js
@@ -1,78 +1,15 @@
 import React from 'react'
-import { Link } from 'gatsby'
-import GithubSlugger from 'github-slugger'
 import glamorous from 'glamorous'
 
-import { TopLevelNavigationItem } from './top-level-navigation-item/'
+import { TopLevelNavigationItem } from './top-level-navigation-item'
+import { NavigationItem } from './NavigationItem'
 
-class MidLevelNavigationItem extends React.Component {
-  state = {
-    folded: true,
-    style: {
-      display: 'flex',
-      flexFlow: 'column nowrap',
-      alignItems: 'flex-start',
-      justifyContent: 'flex-start',
-      maxHeight: 0,
-      overflow: 'hidden',
-      transition: 'max-height 0.2s ease-out'
-    }
-  }
-
-  unfold () {
-    if (this.state.folded) {
-      this.props.onChange && this.props.onChange(this.div.scrollHeight)
-      this.setState({
-        folded: false,
-        style: {
-          ...this.state.style,
-          maxHeight: `${this.div.scrollHeight}px`
-        }
-      })
-    } else {
-      this.props.onChange && this.props.onChange(0)
-      this.setState({
-        folded: true,
-        style: {
-          ...this.state.style,
-          maxHeight: 0
-        }
-      })
-    }
-  }
-
-  setDivRef = element => {
-    this.div = element
-  }
-
-  render () {
-    const { title, path } = this.props
-    return (
-      <div>
-        <Wrapper2 onClick={() => this.unfold()}>
-          <NavigationLink
-            to={path}
-            activeStyle={{
-              color: 'black',
-              fontFamily: 'Roboto Mono, monospace',
-              fontWeight: '500',
-              fontSize: '0.8rem',
-              transition: 'color 0.2s ease-in-out 0s'
-            }}
-            isActive={isActiveGroup}
-          >
-            {title}
-          </NavigationLink>
-        </Wrapper2>
-        <div style={this.state.style} ref={this.setDivRef}>
-          <div style={{marginLeft: '1rem'}}>
-            { this.props.children }
-          </div>
-        </div>
-      </div>
-    )
-  }
-}
+const List = glamorous.ul({
+  listStyle: 'none',
+  paddingTop: '0.5rem',
+  paddingBottom: 0,
+  margin: 0
+})
 
 export class Navigation extends React.PureComponent {
   state = {
@@ -134,102 +71,3 @@ export class Navigation extends React.PureComponent {
     )
   }
 }
-
-const isActiveGroup = (match, location) => {
-  return match && match.isExact
-}
-
-const Wrapper2 = glamorous.div({
-  position: 'relative',
-  display: 'flex',
-  flexFlow: 'row nowrap',
-  alignItems: 'center'
-})
-
-const NavigationItem = ({ node: { headings, frontmatter: { title, path } }, onChange }) => (
-  <li key={path}>
-    <MidLevelNavigationItem path={path} title={title} onChange={onChange}>
-      { headings.length > 0 && <List>
-        { headings.map((heading, index) => NavigationHeading({ ...heading, path, index })) }
-      </List> }
-    </MidLevelNavigationItem>
-  </li>
-)
-
-const NavigationLink = glamorous(Link)(props => ({
-  position: 'relative',
-  left: '1rem',
-  color: 'black',
-  fontFamily: 'Roboto Mono, monospace',
-  fontWeight: '300',
-  fontSize: '0.8rem',
-  textDecoration: 'none',
-  ':hover': {
-    color: 'black'
-  },
-  ':before': {
-    backgroundColor: '#F486CA',
-    content: ' ',
-    position: 'absolute',
-    width: '1px',
-    height: '100%',
-    top: 0,
-    left: '-5px',
-    visibility: 'hidden',
-    transform: 'scaleY(0.3)',
-    transition: 'all 0.3s ease-in-out 0s'
-  },
-  // '.active:before': {
-  //   visibility: 'visible',
-  //   transform: 'scaleY(1)',
-  //   backgroundColor: '#F486CA'
-  // },
-  '.active': {
-    color: 'black',
-    fontFamily: 'Roboto Mono, monospace',
-    fontWeight: '500',
-    fontSize: '0.8rem',
-    transition: 'color 0.2s ease-in-out 0s'
-  },
-  ':hover:before': {
-    visibility: 'visible',
-    transform: 'scaleY(1)'
-  }
-}))
-
-const isActiveSubItem = (match, location, anchor) => {
-  return match && match.isExact && location.hash === `#${anchor}`
-}
-
-const NavigationHeading = ({ path, value, index }) => {
-  const slugger = new GithubSlugger()
-  const anchor = slugger.slug(value)
-  return <ListSubItem key={index}>
-    <NavigationLink
-      to={`${path}#${anchor}`}
-      activeClassName='active'
-      // activeStyle={{
-      //   color: 'black',
-      //   fontFamily: 'Roboto Mono, monospace',
-      //   fontWeight: '500',
-      //   fontSize: '0.8rem',
-      //   transition: 'color 0.2s ease-in-out 0s'
-      // }}
-      isActive={(match, location) => isActiveSubItem(match, location, anchor)}
-    >
-      { value }
-    </NavigationLink>
-  </ListSubItem>
-}
-
-const List = glamorous.ul({
-  listStyle: 'none',
-  // padding: '1rem',
-  paddingTop: '0.5rem',
-  paddingBottom: 0,
-  margin: 0
-})
-
-const ListSubItem = glamorous.li({
-  fontSize: '0.9rem'
-})
