+++ workspaces/homepage/src/components/navigation/mid-level-navigation-item/index.js
@@ -1,0 +1,1 @@
+export { MidLevelNavigationItem } from './MidLevelNavigationItem'
